c#: add workaround for keepalive
- Added a extra task that runs after all the data has been streamed.
This task pings the server for the version every minute in order
to prevent the channel from being closed due to the period of inactivity
between the end of data stream and results coming back from the server.

- This workaround should be removed once keepalive feature is properly
implemented in the C# API for gRPC and can be enabled via ChannelOptions.